Radically simplify the GifSpriteNode
I tried to remove as many class attributes as possible to simplify the class.  Perhaps a bit too radical. ;-)  The duration is now hardcoded because I did not find the duration value in the gif.info was not very helpful.  The commented URL seems to have a pallet issue on one frame.